Re #19: Melody now rendered as volpiano. Generated at work -> line -> pitch level. Not split into syllables, text underlay not added.
